Remove more legacy tab code
